+++ starkeno/esempi/__init__.py
@@ -1,0 +1,1 @@
+"""Blueprint d'esempio spediti col pacchetto."""